requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
dj-database-url==0.5.0
Django==3.2
django-appconf==1.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-googledrive-storage==1.5.2
django-on-heroku==1.1.2
djangorestframework==3.12.4
google-api-core==1.27.0
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
packaging==20.9
# Pillow==8.2.0
protobuf==3.17.0
# psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
waitress==2.0.0
whitenoise==5.2.0
